gpu: nvgpu: Fix MISRA 10.6 violations in nvlink
MISRA rule 10.6 does not allow assigning of composite expression
to an object with wider essential type. Fix 10.6 violations in nvlink
code by changing the data-type or by type-casting.

JIRA NVGPU-1921

Reviewed-on: https://git-master.nvidia.com/r/2022881
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/common/nvlink/intr_and_err_handling_gv100.c
@@ -253,7 +253,8 @@ static void gv100_nvlink_minion_isr(struct gk20a *g) {
 					minion_minion_intr_nonfatal_f(1));
 	}
 
-	links = minion_minion_intr_link_v(intr) & g->nvlink.enabled_links;
+	links = minion_minion_intr_link_v(intr) &
+					(unsigned long) g->nvlink.enabled_links;
 
 	if (links != 0UL) {
 		for_each_set_bit(bit, &links, NVLINK_MAX_LINKS_SW) {
